# Less branchy Udivrem